requirements.txt
asgiref==3.8.1
Django==5.1.4
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg-binary==3.2.3
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3

drf-spectacular~=0.28.0